Refactored and sytlized WanikaniItems control panel
